+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Segment for my own library `partial-recall`.
+- Segment for my own library `partial-recall`. Will toggle implanted
+  state on click.
 
 ## [v0.6.2]